some fixes (PLEASE DON'T UPLOAD NON-COMPILING AND/OR WRONG OR BAD CODE)
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@841 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/server/CServer_Send.cpp
@@ -279,12 +279,12 @@ void GameServer::SendDisconnect(void)
 	
 	bs.writeByte(S2C_LEAVING);
 
-	for(short c=0;c<MAX_CLIENTS;c++,cl++) {
+	for(short c=0; c<MAX_CLIENTS; c++,cl++) {
 		if(cl->getStatus() == NET_DISCONNECTED)
 			continue;
 
 		// Send it out-of-bounds 3 times to make sure all the clients received it
-		for(int i=0;i<3;i++)
+		for(short i=0; i<3; i++)
 			cl->getChannel()->Transmit(&bs);
 	}	
 }
